Add `--type` arg to filter by package type / ecosystem
Filtering PRs by package ecosystem is useful if the Dependabot configuration for
some repos have been changed from the default so that they don't include a label
for the language. Hypothesis has language labels turned off in the client and
browser-extension repos for example, since all updates are the same language.