Add support for Dependabot grouped updates
When this tool encounters a PR that contains a group of updates, it treats the
group name like a package name for the purposes of grouping PRs from different
repositories for review. This is based on the assumption that a team will
typically configure the same or similiar groups across repositories, and so
groups of the same name will likely contain the same updates.

If the same dependency is updated via standalone PRs in some repositories and
groups in other repositories, they will be presented for review separately.

When presenting the details of an update, make it clear whether the update is
for a group or single dependency, and what the version updates for individual
packages within a group are.