Handle grouped updates with only a single update
When a grouped update only contains a single update, the structure of the PR
body is slightly different and unfortunately the from/to version numbers cannot
be parsed out using the same code. For now the from/to versions are simply
missing from the CLI output in that case. The user can probably infer the target
version from the release notes, or if they trust the upstream vendor enough,
just ignore the issue.

Also provide more details about what went wrong if parsing a PR title fails.