Obtain a new GITHUB_TOKEN with more permissions before running a check
The default GITHUB_TOKEN available to workflows does not have permissions to
query Dependabot vulnerabilities organization-wide, hence the tool was finding
no vulnerabilities when it ran. Resolve this by obtaining a new GITHUB_TOKEN
associated with our Hypothesis GitHub App which does have the necessary
permissions. This follows the examples in the cookiecutter repo [1].

[1] https://github.com/hypothesis/cookiecutters/blob/a2aaf018f11ade8d40a04368f772cd23f6f2f9bc/.github/workflows/update_repos.yml#L34-L38